Disables pipx. Need a better path for gnome exts

+++ install/20_gnome-extensions.sh
@@ -5,29 +5,36 @@
 
 printf "Configuring Gnome Extensions...\n\n"
 
-sudo dnf install pipx -y
-pipx install gnome-extensions-cli --system-site-packages
+#pipx not in normal repo.  Need EPEL for it. Removing for now.
+#sudo dnf install pipx -y
+#pipx install gnome-extensions-cli --system-site-packages
 
 # Install new extensions
-gext install tactile@lundal.io
-gext install just-perfection-desktop@just-perfection
+#gext install tactile@lundal.io
+#gext install just-perfection-desktop@just-perfection
 
-#Setup schemas
-sudo cp ~/.local/share/gnome-shell/extensions/tactile@lundal.io/schemas/org.gnome.shell.extensions.tactile.gschema.xml /usr/share/glib-2.0/schemas
-sudo cp ~/.local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/schemas/org.gnome.shell.extensions.just-perfection.gschema.xml /usr/share/glib-2.0/schemas
-sudo glib-compile-schemas /usr/share/glib-2.0/schemas/
+tactile_schema="~/.local/share/gnome-shell/extensions/tactile@lundal.io/schemas/org.gnome.shell.extensions.tactile.gschema.xml"
+justper_schema="~/.local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/schemas/org.gnome.shell.extensions.just-perfection.gschema.xml"
 
-# Tactile Setup:
-gsettings set org.gnome.shell.extensions.tactile col-0 1
-gsettings set org.gnome.shell.extensions.tactile col-1 2
-gsettings set org.gnome.shell.extensions.tactile col-2 1
-gsettings set org.gnome.shell.extensions.tactile col-3 0
-gsettings set org.gnome.shell.extensions.tactile row-0 1
-gsettings set org.gnome.shell.extensions.tactile row-1 1
-gsettings set org.gnome.shell.extensions.tactile gap-size 32
+if [ -f $tactile_schema ] && [ -f $justper_schema ]; then
+	#Setup schemas
+	sudo cp ~/.local/share/gnome-shell/extensions/tactile@lundal.io/schemas/org.gnome.shell.extensions.tactile.gschema.xml /usr/share/glib-2.0/schemas
+	sudo cp ~/.local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/schemas/org.gnome.shell.extensions.just-perfection.gschema.xml /usr/share/glib-2.0/schemas
+	sudo glib-compile-schemas /usr/share/glib-2.0/schemas/
 
-# Configure Just Perfection
-gsettings set org.gnome.shell.extensions.just-perfection animation 2
-gsettings set org.gnome.shell.extensions.just-perfection dash-app-running true
-gsettings set org.gnome.shell.extensions.just-perfection workspace true
-gsettings set org.gnome.shell.extensions.just-perfection workspace-popup false
+	# Tactile Setup:
+	gsettings set org.gnome.shell.extensions.tactile col-0 1
+	gsettings set org.gnome.shell.extensions.tactile col-1 2
+	gsettings set org.gnome.shell.extensions.tactile col-2 1
+	gsettings set org.gnome.shell.extensions.tactile col-3 0
+	gsettings set org.gnome.shell.extensions.tactile row-0 1
+	gsettings set org.gnome.shell.extensions.tactile row-1 1
+	gsettings set org.gnome.shell.extensions.tactile gap-size 32
+
+	# Configure Just Perfection
+	gsettings set org.gnome.shell.extensions.just-perfection animation 2
+	gsettings set org.gnome.shell.extensions.just-perfection dash-app-running true
+	gsettings set org.gnome.shell.extensions.just-perfection workspace true
+	gsettings set org.gnome.shell.extensions.just-perfection workspace-popup false
+else
+	printf "\n\nTactile and Just Perfection not installed\nInstall them first\n\n"
